phase-2/pyproject.toml
[tool.poetry]
name = "phase-2"
version = "0.1.0"
description = "Django Ninja REST API project with unit tests"
authors = ["Developer <dev@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
django-ninja = "^1.2.0"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.12.3"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
faker = "^19.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
